ft 16 + 20
