+++ src/index.html
@@ -12,10 +12,7 @@
     <include src="./partials/header.html"></include>
 
     <main>
-      
       <include src="./partials/modal.html"></include>
-      <button id="myBtn" data-modal-open>Open Modal</button>
-
 
       <include src="./partials/main-container.html"></include>
     </main>
@@ -23,6 +20,6 @@
     <include src="./partials/footer.html"></include>
 
     <script src="./js/header.js" type="module"></script>
-    <script src="./js/modal.js" type="module"></script>
+    <script src="./js/modal-gallery.js" type="module"></script>
   </body>
 </html>
