+++ src/sass/index.scss
@@ -5,6 +5,6 @@
 @import './layot/main-gallery';
 @import './layot/filmcard';
 @import './layot/footer';
-@import "./layot/modal";
+@import "./layot/modal-gallery";
 @import "./layot/backdrop";
 @import './layot/library';
