+++ src/sass/layot/_modal-gallery.scss
