+++ src/templates/modal-gallery.hbs
